Trigger the function that gets the spotify's bearer token when a user logs in or refresh his profile page.
