+++ services/spotify.js
@@ -1,0 +1,22 @@
+//Using Spotify Api
+const CLIENT_ID = process.env.CLIENT_ID;
+const CLIENT_SECRET = process.env.CLIENT_SECRET;
+
+function getApiToken() {
+  return new Promise((resolve, reject) => {
+    var authParameters = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/x-www-form-urlencoded'
+      },
+      body: 'grant_type=client_credentials&client_id=' + CLIENT_ID + '&client_secret=' + CLIENT_SECRET
+    }
+
+    fetch('https://accounts.spotify.com/api/token', authParameters)
+      .then(result => result.json())
+      .then(data => resolve(data.access_token))
+      .catch(error => reject(error));
+  });
+}
+
+module.exports = { getApiToken };
